Fixes for integer width checks.
Turns out that my first implementation of this was flawed. Things like
uint8(uint8(0)) would cause an error.

To make this a bit easier to follow I changed the lexer return tokens for each
of the integer functions, instead of a single _INTEGER_FUNCTION_ token with an
index value to use when emiting the opcode. By doing this I can then set the
width of the expression based upon the new token types and use that when doing
the width check (instead of inferring the width from the index).

Once we have reduced the integer_function we can set it to an integer with an
undefined value and it will be treated like a normal integer for all operations,
except for _EQ_, which has special case logic to check the width and any
compile-time values.