feat: skip bytecode evaluation for some rules without string matches
Optimize a common case where YARA conditions are formed like e.g.
"... and 1 of them and ...", in other words, requiring a string
match to ever be true.
By noting these cases and recording in a bitmap if a string match
occurred, the condition evaluation for these rules can be skipped
entirely in most cases.