Fix inconsistency with `--fast-scan` option.
Some rules were not matching when the `--fast-scan` flag was used, but they should. It happened with rules that contained statements like `any of  <string_set> in <range>` or `any of <string_set> at <offset>`. With this type of expressions, the strings included in `<string_set>` can't be flagged with `STRING_FLAGS_SINGLE_MATCH` because we need to find all the occurrences of those strings. Finding only the first match is not enough because the condition can be true for some other occurrence of the string, but not with the first one..

With this change the flag `STRING_FLAGS_SINGLE_MATCH` is cleared for every string included in a string set. This is a radical way of fixing the issue, as the flag is cleared in other cases where this is not necessary, like in `any of <string_set>`, where finding the first occurrence of each string in the set is enough. But I don't want to add more complexity and correctness should prevail over performance.