fix verified field in pe module with multiple signatures
Commit 1a85738 fixed one bug but introduced a new one. The
`pe.is_signed` field is now properly set (true if any signature is
valid), but the `signatures[i].verified` is now incorrectly set: instead
of being set to whether this signature is verified, it is set to
true if any previous signature is verified.

Unfortunately, there is no test file that is triggering this behavior.
This would require having a signature that is invalid appearing after a
signature that is valid.