Several fixes related to the entrypoint value for the macho module (#1995)
* fix bug in entrypoint computation on process mem scan

When using the PROCESS_MEMORY flag and the macho module, the LC_MAIN
load command was badly handled, the computation of the end of a segment
was invalid.

* add base address to entrypoint when scanning proc memory

As is done for other file analysis modules, the base address of the
scanned region must be added to the entrypoint when scanning with the
PROCESS_MEMORY flag set.

* do not set invalid entrypoint value if command is too small

If the unixthread command is too small, the entrypoint value is not
retrieved. The function still continues executing however, and
considers 0 to be the entrypoint value. This is invalid, and the
entrypoint should stay undefined in those cases.