Fix compiler crash
This fixes a compiler crash while parsing the invalid YARA rules described in #1992. The root of the issue was the incorrect initialization of a `YR_EXPRESSION` structure where the type was set to `EXPRESSION_TYPE_STRING` but the value was initialized as an integer and set to `YR_UNDEFINED`. As the value is an union type, this undefined integer was later interpreted as an arena reference via the `sized_string_ref` union field.

Closes #1992.