Calculate a fully matching regex for base64 strings
The original code threw away up to the first couple of characters, and
the last couple of characters, matching on the portions of the three
permutations that were relatively static. However there is only a limited
subset of characters that those characters could be, and with a little
bit of math and a couple of extra strings in the node we can generate a
regular expression that 100% matches all three permutations of the
string, avoiding false positives on very short strings.